Check directory mtime after added/deleted inodes
Many operations to synchronize inodes alter the modification time of the
parent directory. However, if it is restored to its original value after any
change on source side, fssync could skip synchronization of parent metadata,
resulting in wrong timestamp.

This also fixes a random failure in unit tests because on Linux, granularity of
modification time seems to be the same as CONFIG_HZ, and the test is running
too fast to always produce a different timestamp.